Create a test suite (a stub) for symbol hierarchy
