+++ jvm/src/test/scala/SymbolSuite.scala
@@ -1,0 +1,10 @@
+import tastyquery.Contexts
+import tastyquery.Contexts.Context
+
+class SymbolSuite extends BaseUnpicklingSuite {
+  def getUnpicklingContext(filename: String): Context = {
+    val ctx = Contexts.empty
+    unpickle(filename) (using ctx)
+    ctx
+  }
+}
